extract panel style components
