+++ src/components/GlobalStyle.tsx
@@ -87,7 +87,6 @@ html {
 
 html, body, #__next {
   height: 100%;
-  width: 100%;
 }
 
 body {
