+++ src/components/Panel/ExpiryTypeIcon.tsx
@@ -1,7 +1,7 @@
 import { ExpiryType } from "@/types";
 import Clock from "public/assets/icons/clock.svg";
 import styled from "styled-components";
-import { InfoIconText, InfoIconWrapper } from "./styles";
+import { InfoIconText, InfoIconWrapper } from "./style";
 
 /**
  * Displays an icon for the given expiry type.
