+++ src/components/Panel/OoTypeIcon.tsx
@@ -2,7 +2,7 @@ import { IconWrapper } from "@/components";
 import { OracleType } from "@/types";
 import OptimisticOracle from "public/assets/icons/optimistic-oracle.svg";
 import styled from "styled-components";
-import { InfoIconText, InfoIconWrapper } from "./styles";
+import { InfoIconText, InfoIconWrapper } from "./style";
 
 /**
  * Displays an icon for the given Optimistic Oracle type.
