+++ src/components/Panel/Panel.tsx
@@ -2,7 +2,6 @@ import { Button, DecimalInput } from "@/components";
 import { blueGrey700, currencyIcons, projectIcons, red500 } from "@/constants";
 import { addOpacityToHsl, getValueText } from "@/helpers";
 import { usePanelContext } from "@/hooks";
-import { AnimatePresence, motion } from "framer-motion";
 import NextLink from "next/link";
 import AncillaryData from "public/assets/icons/ancillary-data.svg";
 import Close from "public/assets/icons/close.svg";
@@ -11,15 +10,13 @@ import Pencil from "public/assets/icons/pencil.svg";
 import Settled from "public/assets/icons/settled.svg";
 import Timestamp from "public/assets/icons/timestamp.svg";
 import Warning from "public/assets/icons/warning.svg";
-import { CSSProperties, Fragment, useEffect, useRef, useState } from "react";
-import { FocusOn } from "react-focus-on";
+import { CSSProperties, Fragment, useState } from "react";
 import styled from "styled-components";
+import { Base } from "./Base";
 import { ChainIcon } from "./ChainIcon";
 import { ExpiryTypeIcon } from "./ExpiryTypeIcon";
 import { OoTypeIcon } from "./OoTypeIcon";
 
-const overlayVisibleColor = addOpacityToHsl(blueGrey700, 0.75);
-const overlayHiddenColor = addOpacityToHsl(blueGrey700, 0);
 const errorBackgroundColor = addOpacityToHsl(red500, 0.05);
 
 /**
@@ -30,15 +27,7 @@ const errorBackgroundColor = addOpacityToHsl(red500, 0.05);
  */
 export function Panel() {
   const { content, page, panelOpen, closePanel } = usePanelContext();
-  const contentRef = useRef<HTMLDivElement>(null);
   const [inputValue, setInputValue] = useState("");
-
-  // Scroll to top when panel opens
-  useEffect(() => {
-    if (panelOpen) {
-      contentRef?.current?.scrollTo({ top: 0 });
-    }
-  }, [panelOpen]);
 
   const {
     chainId,
@@ -89,201 +78,139 @@ export function Panel() {
   }
 
   return (
-    <>
-      <AnimatePresence>
-        {panelOpen && (
-          <Overlay
-            onClick={closePanel}
-            initial={{ backgroundColor: overlayHiddenColor }}
-            animate={{ backgroundColor: overlayVisibleColor }}
-            exit={{ backgroundColor: overlayHiddenColor }}
-          />
+    <Base panelOpen={panelOpen} closePanel={closePanel}>
+      <TitleWrapper>
+        <ProjectIconWrapper>{projectIcon}</ProjectIconWrapper>
+        <Title id="panel-title">{title}</Title>
+        <CloseButton aria-label="close panel" onClick={closePanel}>
+          <CloseIconWrapper>
+            <CloseIcon />
+          </CloseIconWrapper>
+        </CloseButton>
+      </TitleWrapper>
+      <ActionsWrapper>
+        <SectionTitleWrapper>
+          {actionsIcon}
+          <SectionTitle>{actionsTitle}</SectionTitle>
+        </SectionTitleWrapper>
+        {hasInput && setError ? (
+          <InputWrapper>
+            <DecimalInput
+              value={inputValue}
+              onInput={setInputValue}
+              disabled={isError}
+              addErrorMessage={setError}
+              removeErrorMessage={() => setError("")}
+            />
+          </InputWrapper>
+        ) : (
+          <ValueWrapper
+            style={
+              {
+                "--margin-bottom": showActionsDetails ? "20px" : "0px",
+              } as CSSProperties
+            }
+          >
+            <ValueText>{valueText}</ValueText>
+          </ValueWrapper>
         )}
-      </AnimatePresence>
-      <FocusOn
-        enabled={panelOpen}
-        onClickOutside={closePanel}
-        onEscapeKey={closePanel}
-        preventScrollOnFocus
-      >
-        <Content
-          ref={contentRef}
-          role="dialog"
-          aria-modal={panelOpen}
-          aria-labelledby="panel-title"
-          style={
-            {
-              "--right": panelOpen ? 0 : "var(--panel-width)",
-            } as CSSProperties
-          }
-        >
-          <TitleWrapper>
-            <ProjectIconWrapper>{projectIcon}</ProjectIconWrapper>
-            <Title id="panel-title">{title}</Title>
-            <CloseButton aria-label="close panel" onClick={closePanel}>
-              <CloseIconWrapper>
-                <CloseIcon />
-              </CloseIconWrapper>
-            </CloseButton>
-          </TitleWrapper>
-          <ActionsWrapper>
-            <SectionTitleWrapper>
-              {actionsIcon}
-              <SectionTitle>{actionsTitle}</SectionTitle>
-            </SectionTitleWrapper>
-            {hasInput && setError ? (
-              <InputWrapper>
-                <DecimalInput
-                  value={inputValue}
-                  onInput={setInputValue}
-                  disabled={isError}
-                  addErrorMessage={setError}
-                  removeErrorMessage={() => setError("")}
-                />
-              </InputWrapper>
-            ) : (
-              <ValueWrapper
-                style={
-                  {
-                    "--margin-bottom": showActionsDetails ? "20px" : "0px",
-                  } as CSSProperties
-                }
-              >
-                <ValueText>{valueText}</ValueText>
-              </ValueWrapper>
-            )}
-            {showActionsDetails && (
-              <ActionsDetailsWrapper>
-                <ActionWrapper>
-                  <ActionText>
-                    Bond
-                    <InfoIcon />
-                  </ActionText>
-                  <ActionText>
-                    {currencyIcon && (
-                      <CurrencyIconWrapper>{currencyIcon}</CurrencyIconWrapper>
-                    )}
-                    {formattedBond} {!currencyIcon && currency}
-                  </ActionText>
-                </ActionWrapper>
-                <ActionWrapper>
-                  <ActionText>
-                    Reward
-                    <InfoIcon />
-                  </ActionText>
-                  <ActionText>
-                    {currencyIcon && (
-                      <CurrencyIconWrapper>{currencyIcon}</CurrencyIconWrapper>
-                    )}
-                    {formattedReward} {!currencyIcon && currency}
-                  </ActionText>
-                </ActionWrapper>
-                <ActionWrapper>
-                  <ActionText>
-                    Challenge period ends in
-                    <InfoIcon />
-                  </ActionText>
-                  <ActionText>{formattedLivenessEndsIn}</ActionText>
-                </ActionWrapper>
-              </ActionsDetailsWrapper>
-            )}
-            {hasActionButton && (
-              <ActionButtonWrapper>
-                <Button
-                  variant="primary"
-                  onClick={action}
-                  width="min(100%, 512px)"
-                >
-                  {actionType}
-                </Button>
-              </ActionButtonWrapper>
-            )}
-            {isError && (
-              <ErrorWrapper>
-                <WarningIcon />
-                <ErrorText>{error}</ErrorText>
-              </ErrorWrapper>
-            )}
-          </ActionsWrapper>
-          <InfoIconsWrapper>
-            <ChainIcon chainId={chainId} />
-            <OoTypeIcon ooType={oracleType} />
-            {expiryType && <ExpiryTypeIcon expiryType={expiryType} />}
-          </InfoIconsWrapper>
-          <DetailsWrapper>
-            <DetailWrapper>
-              <SectionTitleWrapper>
-                <TimestampIcon />
-                <SectionTitle>Timestamp</SectionTitle>
-              </SectionTitleWrapper>
-              <Time>
-                <TimeFormat>UTC</TimeFormat> {timeUTC}
-              </Time>
-              <Time>
-                <TimeFormat>UNIX</TimeFormat> {timeUNIX}
-              </Time>
-            </DetailWrapper>
-            <DetailWrapper>
-              <SectionTitleWrapper>
-                <AncillaryDataIcon />
-                <SectionTitle>Ancillary Data</SectionTitle>
-              </SectionTitleWrapper>
-              <SectionSubTitle>String</SectionSubTitle>
-              <DetailText>{decodedAncillaryData}</DetailText>
-              <SectionSubTitle>Bytes</SectionSubTitle>
-              <DetailText>{ancillaryData}</DetailText>
-            </DetailWrapper>
-            <DetailWrapper>
-              <SectionTitleWrapper>
-                <AncillaryDataIcon />
-                <SectionTitle>More information</SectionTitle>
-              </SectionTitleWrapper>
-              {moreInformation?.map(({ title, href, text }) => (
-                <Fragment key={title}>
-                  <SectionSubTitle>{title}</SectionSubTitle>
-                  <Link href={href} target="_blank">
-                    {text}
-                  </Link>
-                </Fragment>
-              ))}
-            </DetailWrapper>
-          </DetailsWrapper>
-        </Content>
-      </FocusOn>
-    </>
+        {showActionsDetails && (
+          <ActionsDetailsWrapper>
+            <ActionWrapper>
+              <ActionText>
+                Bond
+                <InfoIcon />
+              </ActionText>
+              <ActionText>
+                {currencyIcon && (
+                  <CurrencyIconWrapper>{currencyIcon}</CurrencyIconWrapper>
+                )}
+                {formattedBond} {!currencyIcon && currency}
+              </ActionText>
+            </ActionWrapper>
+            <ActionWrapper>
+              <ActionText>
+                Reward
+                <InfoIcon />
+              </ActionText>
+              <ActionText>
+                {currencyIcon && (
+                  <CurrencyIconWrapper>{currencyIcon}</CurrencyIconWrapper>
+                )}
+                {formattedReward} {!currencyIcon && currency}
+              </ActionText>
+            </ActionWrapper>
+            <ActionWrapper>
+              <ActionText>
+                Challenge period ends in
+                <InfoIcon />
+              </ActionText>
+              <ActionText>{formattedLivenessEndsIn}</ActionText>
+            </ActionWrapper>
+          </ActionsDetailsWrapper>
+        )}
+        {hasActionButton && (
+          <ActionButtonWrapper>
+            <Button variant="primary" onClick={action} width="min(100%, 512px)">
+              {actionType}
+            </Button>
+          </ActionButtonWrapper>
+        )}
+        {isError && (
+          <ErrorWrapper>
+            <WarningIcon />
+            <ErrorText>{error}</ErrorText>
+          </ErrorWrapper>
+        )}
+      </ActionsWrapper>
+      <InfoIconsWrapper>
+        <ChainIcon chainId={chainId} />
+        <OoTypeIcon ooType={oracleType} />
+        {expiryType && <ExpiryTypeIcon expiryType={expiryType} />}
+      </InfoIconsWrapper>
+      <DetailsWrapper>
+        <DetailWrapper>
+          <SectionTitleWrapper>
+            <TimestampIcon />
+            <SectionTitle>Timestamp</SectionTitle>
+          </SectionTitleWrapper>
+          <Time>
+            <TimeFormat>UTC</TimeFormat> {timeUTC}
+          </Time>
+          <Time>
+            <TimeFormat>UNIX</TimeFormat> {timeUNIX}
+          </Time>
+        </DetailWrapper>
+        <DetailWrapper>
+          <SectionTitleWrapper>
+            <AncillaryDataIcon />
+            <SectionTitle>Ancillary Data</SectionTitle>
+          </SectionTitleWrapper>
+          <SectionSubTitle>String</SectionSubTitle>
+          <DetailText>{decodedAncillaryData}</DetailText>
+          <SectionSubTitle>Bytes</SectionSubTitle>
+          <DetailText>{ancillaryData}</DetailText>
+        </DetailWrapper>
+        <DetailWrapper>
+          <SectionTitleWrapper>
+            <AncillaryDataIcon />
+            <SectionTitle>More information</SectionTitle>
+          </SectionTitleWrapper>
+          {moreInformation?.map(({ title, href, text }) => (
+            <Fragment key={title}>
+              <SectionSubTitle>{title}</SectionSubTitle>
+              <Link href={href} target="_blank">
+                {text}
+              </Link>
+            </Fragment>
+          ))}
+        </DetailWrapper>
+      </DetailsWrapper>
+    </Base>
   );
 }
 
 // wrappers
-const AnimatedOverlay = motion.div;
-
-const Overlay = styled(AnimatedOverlay)`
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  width: 100%;
-  height: 100%;
-  overflow-x: hidden;
-  z-index: 1;
-`;
-
-const Content = styled.div`
-  width: var(--panel-width);
-  min-height: 100%;
-  position: fixed;
-  right: 0;
-  top: 0;
-  bottom: 0;
-  margin: 0;
-  padding: 0;
-  transform: translateX(var(--right));
-  background: var(--white);
-  overflow-y: scroll;
-  transition: transform 400ms;
-  z-index: 1;
-`;
 
 const TitleWrapper = styled.div`
   display: flex;
