+++ src/components/Panel/style.tsx
